+++ editor/webapp/game-selector/components/game-selector.vue
@@ -4,19 +4,19 @@
   11.4.2021 KC
 -->
 <template lang="pug">
-#game-selector
-  modal-agb()
-  menu-bar(:elements="menuElements" show-user-box=true)
-  welcome-bar
-  b-container(fluid=true)
-    b-row
-      b-col
-        p.intro Dies ist der Ferropoly Spiel-Editor. Damit kannst Du neue Spiele erstellen oder bestehende bearbeiten. Weitere Infos findest Du auf der&nbsp;
-          a(href='http://www.ferropoly.ch' target='blank') Ferropoly Webseite
-          | .
-    b-row
-      b-col
-        my-games(v-on:gameplays-changed="gameplaysChanged")
+  #game-selector
+    modal-agb()
+    menu-bar(:elements="menuElements" show-user-box=true)
+    welcome-bar(:user-name="userName")
+    b-container(fluid=true)
+      b-row
+        b-col
+          p.intro Dies ist der Ferropoly Spiel-Editor. Damit kannst Du neue Spiele erstellen oder bestehende bearbeiten. Weitere Infos findest Du auf der&nbsp;
+            a(href='http://www.ferropoly.ch' target='blank') Ferropoly Webseite
+            | .
+      b-row
+        b-col
+          my-games(v-on:gameplays-changed="gameplaysChanged")
 </template>
 
 <script>
@@ -24,20 +24,36 @@ import WelcomeBar from './welcome-bar.vue'
 import MyGames from './my-games.vue'
 import ModalAgb from './modal-agb.vue'
 import MenuBar from '../../common/components/menu-bar/menu-bar.vue'
+import {readUserInfo} from "../adapter/userInfo";
 
 
 export default {
-  name      : "game-selector",
+  name      : 'game-selector',
   props     : [],
   data      : function () {
     return {
       gamePlays   : {},
       menuElements: [
-        {title: 'Neues Spiel', href: '/newgame', hide:false}
-      ]
+        {title: 'Neues Spiel', href: '/newgame', hide: false},
+        {title: 'Admin Dashboard', href: '/dashboard', hide: true}
+      ],
+      userName: '',
+      isAdmin: false
     };
   },
-  model     : {},
+  model     : {
+  },
+  created() {
+    let self = this;
+    // Get the User Info
+    readUserInfo((err, info) => {
+      if (!err) {
+        self.userName = info.personalData.forename + ' ' + info.personalData.surname;
+        self.isAdmin = info.roles.admin;
+        self.menuElements[1].hide = !self.isAdmin;
+      }
+    });
+  },
   methods   : {
     /**
      * Event handler when the gameplays changed