+++ editor/webapp/game-selector/components/welcome-bar.vue
@@ -3,18 +3,24 @@
 -->
 <template lang="pug">
 div
-  #info-header(v-bind:style="{backgroundImage: backgroundImage}")
+  #info-header(:style="{backgroundImage: backgroundImage}")
     #info-welcome
-      h1.info-header {{welcomeText}}
+      h1.info-header {{greetingText}}
 </template>
 
 <script>
 import {DateTime} from "luxon";
-import {readUserInfo} from "../adapter/userInfo";
 
 export default {
   name : "welcome-bar",
-  props: [],
+  props: {
+    userName: {
+      type: String,
+      default: () => {
+        return '';
+      }
+    }
+  },
   data : function () {
     return {
       welcomeText    : '',
@@ -26,32 +32,29 @@ export default {
    * Called when component is being created
    */
   created   : function () {
-    let self             = this;
-
     // Define Background
     let index            = DateTime.now().minute % 10;
     this.backgroundImage = `url("/images/ferropoly_header_0${index}.jpg")`;
-
-    // Be kind and say hello
-    let currentHour      = DateTime.now().hour;
-    let greeting         = 'Hallo'
-    if (currentHour < 4) {
-      greeting = 'Hallo';
-    } else if (currentHour < 10) {
-      greeting = 'Guten Morgen';
-    } else if (currentHour < 17) {
-      greeting = 'Hallo';
-    } else {
-      greeting = 'Guten Abend';
-    }
-    // Get the User Info
-    readUserInfo((err, info) => {
-      if (!err) {
-        self.welcomeText = greeting + ' ' + info.personalData.forename + ' ' + info.personalData.surname
-      }
-    });
   },
   methods   : {},
+  computed: {
+    greetingText() {
+      // Be kind and say hello
+      let currentHour      = DateTime.now().hour;
+      let greeting         = 'Hallo'
+      if (currentHour < 4) {
+        greeting = 'Hallo';
+      } else if (currentHour < 10) {
+        greeting = 'Guten Morgen';
+      } else if (currentHour < 17) {
+        greeting = 'Hallo';
+      } else {
+        greeting = 'Guten Abend';
+      }
+      greeting += ' ' + this.userName;
+      return greeting;
+    }
+  },
   components: {}
 }
 </script>