+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Neuer Editor für Spielregeln
 * Neue Einstellungen für Chance/Kanzlei möglich
 * Neue Bilder für Login und Startseite
+* Dashboard für Admins
 
 # V2.6.1 19.12.2020
 * API für Abfragen aus Webseite und Überwachung #41